fix stack navigations
