+++ src/views/ArticleView/component/NewsArticleComponent.tsx
@@ -4,7 +4,7 @@ import { colors } from '@styles';
 import { NewsArticleItem } from '../types';
 import { dateTimeUtils } from '@utils';
 import { IconType } from '@components';
-import ArticleLayout from '@components/layers/articleLayout/ArticleLayout';
+import ArticleLayout from '@components/layouts/articleLayout/ArticleLayout';
 import { IconBtn44 } from '@components/buttons';
 import TextColorFilledBtn from '@components/buttons/TextColorFilledBtn';
 import translations from '@localization/translations';
@@ -19,35 +19,33 @@ interface Props {
 }
 
 const NewsArticleComponent: React.FC<Props> = ({ item, onBack, onFavourite, onShare, loading, buyTicket }) => {
-
-    const { title, image, content } = item
-    const secondaryTitle = `Ievietots ${dateTimeUtils.formatDate(item.date)}`
+    const { title, image, content } = item;
+    const secondaryTitle = `Ievietots ${dateTimeUtils.formatDate(item.date)}`;
 
     return (
-        <ArticleLayout
-            {...{ onBack, loading, title, image, content, secondaryTitle }}
-        >
+        <ArticleLayout {...{ onBack, loading, title, image, content, secondaryTitle }}>
             <View style={styles.row}>
-                <IconBtn44 onPress={onShare}
+                <IconBtn44
+                    onPress={onShare}
                     style={styles.roundedButton}
                     icon={IconType.Share}
                     color={colors.white}
-                    bgColor={colors.blue} />
+                    bgColor={colors.blue}
+                />
 
-                <IconBtn44 onPress={onFavourite}
+                <IconBtn44
+                    onPress={onFavourite}
                     style={styles.roundedButton}
                     icon={item.isFavourite ? IconType.Heart : IconType.HeartFilled}
                     color={colors.white}
-                    bgColor={colors.orange} />
+                    bgColor={colors.orange}
+                />
 
-                {buyTicket &&
-                    <TextColorFilledBtn
-                        style={styles.buyTicket}
-                        onPress={buyTicket}
-                    >
+                {buyTicket && (
+                    <TextColorFilledBtn style={styles.buyTicket} onPress={buyTicket}>
                         {translations.getString('BUY_TICKET')}
                     </TextColorFilledBtn>
-                }
+                )}
             </View>
         </ArticleLayout>
     );
@@ -56,6 +54,7 @@ const NewsArticleComponent: React.FC<Props> = ({ item, onBack, onFavourite, onSh
 const styles = StyleSheet.create({
     row: {
         flexDirection: 'row',
+        alignItems: 'center',
     },
     roundedButton: {
         borderRadius: 3,
@@ -63,13 +62,8 @@ const styles = StyleSheet.create({
         marginRight: 16,
     },
     buyTicket: {
-        borderRadius: 3,
-        overflow: 'hidden',
-        marginRight: 16,
-        backgroundColor: colors.yellow
-    }
-})
-
-
+        backgroundColor: colors.yellow,
+    },
+});
 
 export default NewsArticleComponent;
